requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1

# Database
motor==3.4.0
pymongo==4.7.2

# AI/ML
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0
tqdm>=4.65.0
scipy>=1.10.0

# Utils
aiofiles==23.2.1
